Feature add: Product search Admin

+++ src/main/java/com/dcava/dcava_backend/repository/UserAdminRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserAdminRepository extends JpaRepository<UserAdmin, Long> {
+public interface UserAdminRepository extends JpaRepository<UserAdmin, Integer> {
     Optional<UserAdmin> findByUidFirebase(String uidFirebase);
 }
